Change protocol names and add the to new files
