+++ MercadoPagoSDK/MercadoPagoSDK/Flows/One tap/PXOneTapResultHandlerProtocol.swift
@@ -1,0 +1,16 @@
+//
+//  PXOneTapResultHandlerProtocol.swift
+//  MercadoPagoSDK
+//
+//  Created by Eden Torres on 03/07/2018.
+//  Copyright © 2018 MercadoPago. All rights reserved.
+//
+
+import Foundation
+protocol PXOneTapResultHandlerProtocol: NSObjectProtocol {
+    func finishOneTap(paymentResult: PaymentResult)
+    func finishOneTap(businessResult: PXBusinessResult)
+    func finishOneTap(paymentData: PaymentData)
+    func cancelOneTap()
+    func exitCheckout()
+}
