+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoCheckoutScreens.swift
@@ -324,42 +324,3 @@ extension MercadoPagoCheckout {
         onetapFlow.start()
     }
 }
-
-extension MercadoPagoCheckout: PXOneTapResultHandler {
-    func finishOneTap(paymentData: PaymentData) {
-        self.viewModel.updateCheckoutModel(paymentData: paymentData)
-        self.executeNextStep()
-    }
-
-    func cancelOneTap() {
-        self.viewModel.prepareForNewSelection()
-        self.executeNextStep()
-    }
-
-    func exitCheckout() {
-        self.finish()
-    }
-
-    func finishOneTap(paymentResult: PaymentResult) {
-        self.setPaymentResult(paymentResult: paymentResult)
-        self.executeNextStep()
-    }
-
-    func finishOneTap(businessResult: PXBusinessResult) {
-        self.viewModel.businessResult = businessResult
-        self.executeNextStep()
-    }
-}
-
-protocol PXOneTapResultHandler: NSObjectProtocol {
-    func finishOneTap(paymentResult: PaymentResult)
-    func finishOneTap(businessResult: PXBusinessResult)
-    func finishOneTap(paymentData: PaymentData)
-    func cancelOneTap()
-    func exitCheckout()
-}
-
-protocol PXPaymentResultHandler: NSObjectProtocol {
-    func finishPaymentFlow(paymentResult: PaymentResult)
-    func finishPaymentFlow(businessResult: PXBusinessResult)
-}
