+++ MercadoPagoSDK/MercadoPagoSDK/PaymentMethodPlugins/PXPluginComponent.swift
@@ -15,8 +15,8 @@ import Foundation
 
 @objc public protocol PXPaymentPluginComponent: PXPluginComponent {
     @objc optional func support(pluginStore: PXCheckoutStore) -> Bool
-    @objc optional func createPaymentWithBusinessResult(pluginStore: PXCheckoutStore, handler: PXPaymentFlowHandler) -> PXBusinessResult
-    @objc optional func createPayment(pluginStore: PXCheckoutStore, handler: PXPaymentFlowHandler) -> PXPaymentPluginResult
+    @objc optional func createPaymentWithBusinessResult(pluginStore: PXCheckoutStore, handler: PXPaymentFlowHandlerProtocol) -> PXBusinessResult
+    @objc optional func createPayment(pluginStore: PXCheckoutStore, handler: PXPaymentFlowHandlerProtocol) -> PXPaymentPluginResult
 }
 
 @objc public protocol PXPluginComponent: PXCustomComponentizable {
